properly finish refresh loop
taskler should in no way run `process.exit` because you don't know if the main program still wants to do something. e.g. the execution of the `done` callback could be async. But the process would exit before this would be properly finished then.

Additionally exiting with error code 1 suggested that there was an error which is not the case!

(and I converted all tabs to spaces. they were mixed :-( )